Moved pokemon sub class implementations to separate file
