+++ pokemonabc.cpp
@@ -7,7 +7,6 @@ Term:	    Fall 2020
 
 This file contains the implementations for classes Pokemon, Pikachu, Charmander,
 Squirtle, and Bulbasaur.  Pokemon is an ABC class.
-Function information below.
 */
 
 #include "pokemon.h"
@@ -29,9 +28,8 @@ Pokemon::Pokemon(AddOnsDb *db, string aName) : type(normal), status(alive), grow
 }
 
 // copy constructor
-Pokemon::Pokemon(const Pokemon &source) : type(source.type), status(source.status), grown(source.grown), level(source.level), exp(source.exp), hp(source.hp), maxHP(source.maxHP), name(source.name), database(0), moves(0), lcolor(BLACK), rcolor(BLACK), left(0), right(0)
+Pokemon::Pokemon(const Pokemon &source) : type(source.type), status(source.status), grown(source.grown), level(source.level), exp(source.exp), hp(source.hp), maxHP(source.maxHP), name(source.name), database(source.database), moves(0), lcolor(BLACK), rcolor(BLACK), left(0), right(0)
 {
-	database = source.database;
 	moves = new AddOns *[MAX_MOVES];
 	for (int i{0}; i < MAX_MOVES; ++i)
 		moves[i] = nullptr;
@@ -42,7 +40,7 @@ Pokemon::Pokemon(const Pokemon &source) : type(source.type), status(source.statu
 Pokemon::~Pokemon()
 {
 	database = NULL;
-	destroy();
+	destroyMovesList();
 	delete[] moves;
 	moves = NULL;
 }
@@ -54,16 +52,12 @@ void Pokemon::copy(AddOns **source)
 		moves[ITEM] = new Items(*(dynamic_cast<Items *>(source[ITEM])));
 
 	for (int i{1}; i < MAX_MOVES; ++i)
-	{
 		if (source[i])
-		{
 			moves[i] = new Attacks(*(dynamic_cast<Attacks *>(source[i])));
-		}
-	}
 }
 
-// destroy table
-void Pokemon::destroy()
+// destroy moves list
+void Pokemon::destroyMovesList()
 {
 	for (int i{0}; i < MAX_MOVES; ++i)
 		delete moves[i];
@@ -83,7 +77,7 @@ int Pokemon::displayMoves() const
 }
 
 // display pokemon full information
-void Pokemon::fullInfo() const
+void Pokemon::displayFullInfo() const
 {
 	char types[][10] = {"electric", "fire", "water", "grass", "normal"};
 	char state[][10] = {"alive", "knock-out"};
@@ -91,7 +85,7 @@ void Pokemon::fullInfo() const
 	cout << "Level: " << level << endl;
 	cout << "Status: " << state[status] << endl;
 	cout << "Hp: " << hp << endl;
-	if (holding())
+	if (isHoldingItem())
 	{
 		cout << name << " is holding a " << *moves[ITEM] << endl;
 	}
@@ -104,15 +98,15 @@ void Pokemon::fullInfo() const
 }
 
 // display only pokemon battle stats
-void Pokemon::battleStats() const
+void Pokemon::displayBattleStats() const
 {
 	char state[][10] = {"alive", "knock-out"};
 	cout << name << "'s current stats-->level: " << level << ", status: " << state[status] << ", hp: " << hp;
 	cout << endl;
 }
 
-// learn helper function
-bool Pokemon::learn()
+// learn new move
+bool Pokemon::learnNewAttack()
 {
 	char translate[][10] = {"electric", "fire", "water", "grass", "normal"};
 	string mtype = translate[type];
@@ -162,44 +156,41 @@ bool Pokemon::learn()
 // return false if user decides not to switch
 bool Pokemon::switchAttacks(AddOns *attack)
 {
-	bool done = false;
-	cout << "0) go back\n";
+	cout << "0. go back\n";
 	displayMoves();
 	int select = minalib::getValidateInt("Select a move to replace: ", 0, MOVES);
 	cout << "You selected " << select << endl;
-	if (select > 0)
-	{
-		delete moves[select];
-		moves[select] = attack;
-		done = true;
-	}
-	return done;
+	if (select == 0)
+		return false;
+
+	delete moves[select];
+	moves[select] = attack;
+
+	return true;
 }
 
 // initializes pokemon with level, hp, and starting move
 bool Pokemon::initialize()
 {
-	bool done = false;
-	if (database)
-	{
-		moves[1] = new Attacks(*(dynamic_cast<Attacks *>(database->retrieve("normal"))));
-		++level;
-		hp += 500;
-		maxHP += 500;
-		done = true;
-	}
-	else
+	if (!database)
 	{
 		cout << "Must add database\n";
+		return false;
 	}
-	return done;
+
+	moves[1] = new Attacks(*(dynamic_cast<Attacks *>(database->retrieve("normal"))));
+	++level;
+	hp += 500;
+	maxHP += 500;
+
+	return true;
 }
 
 // add new item to moves set else indicate pokemon is holding an item already
 bool Pokemon::hold(const Items &item)
 {
 	bool take = false;
-	if (!holding())
+	if (!isHoldingItem())
 	{
 		moves[ITEM] = new Items(item);
 		take = true;
@@ -219,7 +210,7 @@ Pokemon &Pokemon::operator+=(const Items &item)
 }
 
 // indicate pokemon is holding an item or not
-bool Pokemon::holding() const
+bool Pokemon::isHoldingItem() const
 {
 	return moves[ITEM] != NULL;
 }
@@ -295,22 +286,23 @@ void Pokemon::changeName(const string &newName)
 //= operator overloading
 Pokemon &Pokemon::operator=(Pokemon &source)
 {
-	if (&source != this)
-	{
-		type = source.type;
-		status = source.status;
-		level = source.level;
-		exp = source.exp;
-		hp = source.hp;
-		maxHP = source.maxHP;
-		name = source.name;
-		database = source.database;
+	if (&source == this)
+		return *this;
 
-		destroy();
-		delete[] moves;
-		moves = new AddOns *[MAX_MOVES];
-		copy(source.moves);
-	}
+	type = source.type;
+	status = source.status;
+	level = source.level;
+	exp = source.exp;
+	hp = source.hp;
+	maxHP = source.maxHP;
+	name = source.name;
+	database = source.database;
+
+	destroyMovesList();
+	delete[] moves;
+	moves = new AddOns *[MAX_MOVES];
+	copy(source.moves);
+
 	return *this;
 }
 
@@ -328,16 +320,19 @@ istream &operator>>(istream &is, Pokemon &pokemon)
 	return is;
 }
 
+//+ operator overload
 string operator+(const string &str, const Pokemon &pokemon)
 {
 	return str + pokemon.name;
 }
 
+//+ operator overload
 string operator+(const Pokemon &pokemon, const string &str)
 {
 	return pokemon.name + str;
 }
 
+//+ operator overload
 string operator+(const Pokemon &pokemon1, const Pokemon &pokemon2)
 {
 	return pokemon1.name + pokemon2.name;
@@ -473,370 +468,4 @@ bool &Pokemon::leftColor()
 bool &Pokemon::rightColor()
 {
 	return rcolor;
-}
-
-// default constructor
-Pikachu::Pikachu() : factor(0)
-{
-	factor = new float[TYPES]{1.0, 2.0, 2.0, 1.0, 1.0};
-	type = electric;
-}
-
-// arg constructor
-Pikachu::Pikachu(AddOnsDb *db, string aName) : Pokemon(db, aName), factor(0)
-{
-	factor = new float[TYPES]{1.0, 2.0, 2.0, 1.0, 1.0};
-	type = electric;
-}
-
-// copy constructor
-Pikachu::Pikachu(const Pikachu &source) : Pokemon(source), factor(0)
-{
-	factor = new float[source.TYPES];
-	for (int i = 0; i < source.TYPES; ++i)
-	{
-		factor[i] = source.factor[i];
-	}
-	type = source.type;
-}
-
-// destructor
-Pikachu::~Pikachu()
-{
-	delete[] factor;
-	factor = NULL;
-}
-
-// display pikachu name
-// had more to this before ideas changed
-void Pikachu::display(ostream &os) const
-{
-	os << name;
-}
-
-// display pikachu's full information
-void Pikachu::fullInfo() const
-{
-	cout << "Your Pikachu's name is: " << name << endl;
-	Pokemon::fullInfo();
-}
-
-// takes damage an attack, use item if holding item, and change status if hp reaches 0
-bool Pikachu::hit(const Attacks &attack)
-{
-	bool hurt = false;
-	int damage = attack.calcDamage(factor, type);
-	if (damage)
-	{
-		hp -= damage;
-		hurt = true;
-	}
-	else
-	{
-		cout << "Hooray! " << name << " dodged the attack!\n";
-	}
-	if (hp < 100 && hp > 0 && holding())
-	{
-		hp += moves[ITEM]->use();
-		cout << name << " used " << *moves[ITEM] << " to increase its hp by a little!\n";
-		delete moves[ITEM];
-		moves[ITEM] = NULL;
-	}
-	if (hp <= 0)
-	{
-		hp = 0;
-		status = ko;
-	}
-	return hurt;
-}
-
-//= operator overload
-Pikachu &Pikachu::operator=(Pikachu &source)
-{
-	if (&source != this)
-	{
-		Pokemon::operator=(source);
-		delete[] factor;
-		factor = new float[source.TYPES];
-		for (int i = 0; i < source.TYPES; ++i)
-		{
-			factor[i] = source.factor[i];
-		}
-		type = source.type;
-	}
-	return *this;
-}
-
-// default constructor
-Charmander::Charmander() : factor(0)
-{
-	factor = new float[TYPES]{2.0, 1.0, 0.5, 2.5, 1.0};
-	type = fire;
-}
-
-// arg constructor
-Charmander::Charmander(AddOnsDb *db, string aName) : Pokemon(db, aName), factor(0)
-{
-	factor = new float[TYPES]{2.0, 1.0, 0.5, 2.5, 1.0};
-	type = fire;
-}
-
-// copy constructor
-Charmander::Charmander(const Charmander &source) : Pokemon(source), factor(0)
-{
-	factor = new float[source.TYPES];
-	for (int i = 0; i < source.TYPES; ++i)
-	{
-		factor[i] = source.factor[i];
-	}
-	type = source.type;
-}
-
-// destructor
-Charmander::~Charmander()
-{
-	delete[] factor;
-	factor = NULL;
-}
-
-// display charmander's name
-void Charmander::display(ostream &os) const
-{
-	os << name;
-}
-
-// display charmander's full info
-void Charmander::fullInfo() const
-{
-	cout << "Your Charmander's name is: " << name << endl;
-	Pokemon::fullInfo();
-}
-
-// take damage from an attack, use item if holding, change status if knocked out
-bool Charmander::hit(const Attacks &attack)
-{
-	bool hurt = false;
-	int damage = attack.calcDamage(factor, type);
-	if (damage)
-	{
-		hp -= damage;
-		hurt = true;
-	}
-	else
-	{
-		cout << "Hooray! " << name << " dodged the attack!\n";
-	}
-	if (hp < 100 && hp > 0 && holding())
-	{
-		hp += moves[ITEM]->use();
-		cout << name << " used " << *moves[ITEM] << " to increase its hp by a little!\n";
-		delete moves[ITEM];
-		moves[ITEM] = NULL;
-	}
-	if (hp <= 0)
-	{
-		hp = 0;
-		status = ko;
-	}
-	return hurt;
-}
-
-//= operator overload
-Charmander &Charmander::operator=(Charmander &source)
-{
-	if (&source != this)
-	{
-		Pokemon::operator=(source);
-		delete[] factor;
-		factor = new float[source.TYPES];
-		for (int i = 0; i < source.TYPES; ++i)
-		{
-			factor[i] = source.factor[i];
-		}
-		type = source.type;
-	}
-	return *this;
-}
-
-// default constructor
-Squirtle::Squirtle() : factor(0)
-{
-	factor = new float[TYPES]{2.0, 2.5, 1.0, 0.5, 1.0};
-	type = water;
-}
-
-// arg constructor
-Squirtle::Squirtle(AddOnsDb *db, string aName) : Pokemon(db, aName), factor(0)
-{
-	factor = new float[TYPES]{2.0, 2.5, 1.0, 0.5, 1.0};
-	type = water;
-}
-
-// copy constructor
-Squirtle::Squirtle(const Squirtle &source) : Pokemon(source), factor(0)
-{
-	factor = new float[source.TYPES];
-	for (int i = 0; i < source.TYPES; ++i)
-	{
-		factor[i] = source.factor[i];
-	}
-	type = source.type;
-}
-
-// destructor
-Squirtle::~Squirtle()
-{
-	delete[] factor;
-	factor = NULL;
-}
-
-// display squirtle's name
-void Squirtle::display(ostream &os) const
-{
-	os << name;
-}
-
-// display full info
-void Squirtle::fullInfo() const
-{
-	cout << "Your Squirtle's name is: " << name << endl;
-	Pokemon::fullInfo();
-}
-
-// take damage from an attack
-bool Squirtle::hit(const Attacks &attack)
-{
-	bool hurt = false;
-	int damage = attack.calcDamage(factor, type);
-	if (damage)
-	{
-		hp -= damage;
-		hurt = true;
-	}
-	else
-	{
-		cout << "Hooray! " << name << " dodged the attack!\n";
-	}
-	if (hp < 100 && hp > 0 && holding())
-	{
-		hp += moves[ITEM]->use();
-		cout << name << " used " << *moves[ITEM] << " to increase its hp by a little!\n";
-		delete moves[ITEM];
-		moves[ITEM] = NULL;
-	}
-	if (hp <= 0)
-	{
-		hp = 0;
-		status = ko;
-	}
-	return hurt;
-}
-
-//= operator overload
-Squirtle &Squirtle::operator=(Squirtle &source)
-{
-	if (&source != this)
-	{
-		Pokemon::operator=(source);
-		delete[] factor;
-		factor = new float[source.TYPES];
-		for (int i = 0; i < source.TYPES; ++i)
-		{
-			factor[i] = source.factor[i];
-		}
-		type = source.type;
-	}
-	return *this;
-}
-
-// default constructor
-Bulbasaur::Bulbasaur() : factor(0)
-{
-	factor = new float[TYPES]{1.0, 0.5, 2.5, 1.0, 1.0};
-	type = grass;
-}
-
-// arg constructor
-Bulbasaur::Bulbasaur(AddOnsDb *db, string aName) : Pokemon(db, aName), factor(0)
-{
-	factor = new float[TYPES]{1.0, 0.5, 2.5, 1.0, 1.0};
-	type = grass;
-}
-
-// copy construcotr
-Bulbasaur::Bulbasaur(const Bulbasaur &source) : Pokemon(source), factor(0)
-{
-	factor = new float[source.TYPES];
-	for (int i = 0; i < source.TYPES; ++i)
-	{
-		factor[i] = source.factor[i];
-	}
-	type = source.type;
-}
-
-// destructor
-Bulbasaur::~Bulbasaur()
-{
-	delete[] factor;
-	factor = NULL;
-}
-
-// display Bulbasaur's name
-void Bulbasaur::display(ostream &os) const
-{
-	os << name;
-}
-
-// display full information
-void Bulbasaur::fullInfo() const
-{
-	cout << "Your Bulbasaur's name is: " << name << endl;
-	Pokemon::fullInfo();
-}
-
-// take damage from an attack
-// these were placed in the derived class because of an early idea that didn't pan out
-bool Bulbasaur::hit(const Attacks &attack)
-{
-	bool hurt = false;
-	int damage = attack.calcDamage(factor, type);
-	if (damage)
-	{
-		hp -= damage;
-		hurt = true;
-	}
-	else
-	{
-		cout << "Hooray! " << name << " dodged the attack!\n";
-	}
-	if (hp < 100 && hp > 0 && holding())
-	{
-		hp += moves[ITEM]->use();
-		cout << name << " used " << *moves[ITEM] << " to increase its hp by a little!\n";
-		delete moves[ITEM];
-		moves[ITEM] = NULL;
-	}
-	if (hp <= 0)
-	{
-		hp = 0;
-		status = ko;
-	}
-	return hurt;
-}
-
-//= operator overload
-Bulbasaur &Bulbasaur::operator=(Bulbasaur &source)
-{
-	if (&source != this)
-	{
-		Pokemon::operator=(source);
-		delete[] factor;
-		factor = new float[source.TYPES];
-		for (int i = 0; i < source.TYPES; ++i)
-		{
-			factor[i] = source.factor[i];
-		}
-		type = source.type;
-	}
-	return *this;
 }
